+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Fixes [#93](https://github.com/danm-de/Fractions/issues/93): DecimalNotationFormatter throws FormatException for some custom formats (via [Pull-Request](https://github.com/danm-de/Fractions/pulls)) by [lipchev](https://github.com/lipchev)
 - Fixes [#92](https://github.com/danm-de/Fractions/issues/92): Mark the MathExt exensions as obsolete (reported by [lipchev](https://github.com/lipchev))
+- Fixes Fraction.Sqrt(): When the root was calculated from a fraction with a very large numerator and denominator the result could have exceeded double's number range and returned NaN.
 
 ## 8.1.0